requirements.txt
google-api-python-client>=2.190.0
google-auth-httplib2>=0.3.0
google-auth-oauthlib>=1.2.4
rich>=13.0.0
openai>=1.0.0
html2text>=2020.1.16
claude-agent-sdk>=0.1.44
openai-agents>=0.10.2
fastapi>=0.115.0
uvicorn>=0.34.0
python-dotenv>=1.0.0
markitdown[pdf]>=0.1.0
python-telegram-bot>=21.0
